pyproject.toml
[tool.poetry]
name = "soschu-temp"
version = "0.1.0"
description = "Outil d'ajustement des températures dans des fichiers météo basé sur l'irradiation solaire"
authors = ["Laurent Bihin <laurent.bihin@gmail.com>"]
license = "AGPL-3.0 license"
readme = "README.md"
package-mode = true
packages = [{ include = "src" }, { include = "tools" }]

[tool.poetry.dependencies]
python =">=3.9,<3.14"
lxml = "^5.3.0"
pytz = "^2024.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pillow = "^11.2.1"
pyinstaller = "^6.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
build-exe = "tools.build_exe:main"
